+++ src/datagrid/column-service/column-types/column-types.js
@@ -5,3 +5,4 @@ export BooleanType from './boolean-renderers';
 export CheckboxType from './checkbox-renderers';
 export CurrencyType from './currency-renderers';
 export PrimitiveType from './primitive-renderers';
+export MultiSelectType from './multiselect-renderers';