+++ src_docs/components/datagrid/datagrid.constants.jsx
@@ -38,9 +38,13 @@ export const selectTranslations = {
   noResultsText: 'No hits found',
 };
 
-export const countryOptions = countries.all.map(country => (
-  { value: country.alpha3, label: country.name }
-));
+export const countryOptions = countries.all
+  .filter(country => country.alpha3)
+  .map(country => ({ value: country.alpha3, label: country.name }));
+
+export const currencyOptions = countries.all
+  .filter(country => country.alpha3 && country.currencies && country.currencies.length >= 1)
+  .map(country => ({ value: country.alpha3, label: country.currencies[0] }));
 
 export const columns = [
   {
@@ -77,11 +81,22 @@ export const columns = [
     header: 'Country',
     valueKeyPath: ['country'],
     valueType: 'text',
-    componentType: 'select',
+    componentType: 'multiselect',
     selectComponentOptions: countryOptions,
     valueRender: (data) => {
       const country = countryOptions.find(v => v.value === data.get('country'));
       return country ? country.label : null;
+    },
+  },
+  {
+    header: 'Currency',
+    valueKeyPath: ['currency'],
+    valueType: 'text',
+    componentType: 'select',
+    selectComponentOptions: currencyOptions,
+    valueRender: (data) => {
+      const currency = currencyOptions.find(v => v.value === data.get('country'));
+      return currency ? currency.label : null;
     },
   },
   {