+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added multi-select filter
 * Added pagination footer and changed requesting/sorting/filtering of paginated data
+* Fixed currency&float filtering
 
 ## 7.3.14
 * Fixed filtering data in paginated grid